Refactored share module classes and minor fix
