+++ share/src/main/java/org/apache/kafka/server/share/context/FinalContext.java
@@ -15,13 +15,14 @@
  * limitations under the License.
  */
 
-package org.apache.kafka.server.share;
+package org.apache.kafka.server.share.context;
 
 import org.apache.kafka.common.TopicIdPartition;
 import org.apache.kafka.common.Uuid;
 import org.apache.kafka.common.message.ShareFetchResponseData;
 import org.apache.kafka.common.protocol.Errors;
 import org.apache.kafka.common.requests.ShareFetchResponse;
+import org.apache.kafka.server.share.ErroneousAndValidPartitionData;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
