+++ share/src/main/java/org/apache/kafka/server/share/session/ShareSessionKey.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.kafka.server.share;
+package org.apache.kafka.server.share.session;
 
 import org.apache.kafka.common.Uuid;
