+++ share/src/test/java/org/apache/kafka/server/share/session/ShareSessionTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.server.share;
+package org.apache.kafka.server.share.session;
 
 import org.apache.kafka.common.TopicIdPartition;
 import org.apache.kafka.common.Uuid;
